+++ packages/client/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+
+## [0.3.x] - 2025-05-22
+
+### Added
+- Bulk Transfer module improvements:
+  - Added validation to prevent negative transfer amounts
+  - Added real-time Sui address validation with visual feedback
+
+
 ## [0.3.0] - 2025-05-16
 
 ### Added